refactor: rename general metadata directory
