+++ src/component/generalMetadata/generalMetadata.scss
